MBS-9871: Display all useful series relationships in RelatedSeries (#1941)
The only thing we want to avoid here is showing all the part-of-series
rels on the page for every entity in a series. But for that there is
no need to limit the rels shown to URLs - we have a perfectly effective
isNotSeriesPart function we already use for series pages.