MBS-11420: "Paste Credits" doesn't apply credited name (#1960)
The basic flow leading to the issue is this:

  * `this.setState` in `pasteArtistCredit` is called.
  * `componentDidUpdate` is triggered on the artist Autocomplete.
  * Since the artist changed, this sets the observable
    `autocomplete.currentSelection` to the new artist. This observable
    has a subscription, `this.props.onChange`, which updates the artist
    again on the parent (even though it already knows about this
    change).

The bit that screws up the credited name is the last point, because when
the "duplicate" update winds up in `handleArtistChange`, it doesn't know
about the credited name from the original update -- it thinks we're
changing the artist only, so acts to keep the credited name in sync as
it usually does.

We can avoid this by not re-attaching the observable subscription until
after we update the observable. I cannot think of any reason this wasn't
done before, other than oversight.