MBS-10327: Enable subscribing to private collection you collaborate on
If you are able to participate in a private collection, you should also
be able to subscribe to it. Of course, if you're then removed as a
collaborator (thus losing the right to see the collection)
then you should also be removed as a subscriber.

This requires a schema change for the del_collection_sub_on_private
function, since otherwise the collection subscription will be
deactivated for collaborators if the collection was public
and is then made private.