MBS-8820: Show explanation if artist relationships tab shows no rels
For an artist that only has event relationships, the relationships tab
is currently just empty.
This changes it so that Relationships displays an informative message
if all existing relationships have been filtered out.
While at it, I also got rid of the separate "no rels" message
on ArtistRelationships and LabelRelationships, and moved that
to Relationships as well.