Standardise building display entity in ::Create edits
We did this in, if I counted right, five slightly different ways.
Some didn't check for the entity_id at all before trying to use it
(possibly triggering warnings about using an undef value as a hash key),
some did, but only whether it was truthy, and some did and checked if
it's defined.
For extra safety, this always checks whether it's defined now.
This stops passing the entity ID to Instrument->new, since we should
still display something even if the id is missing anyway.
Once we can generate types from the real edit data and are certain
that a specific edit type always has an entity id, we might be able
to improve how we do all this stuff (hopefully on the JS side).