MBS-11454: Don't consider hasTooEarlyFormat a blocking error
If a release already had a too early format before the editor started
this release editor session, hasTooEarlyFormat is still set. As such,
these lines were blocking editing even in that case.
The whole point of having hasUnconfirmedEarlyFormat as well was for
that to be the one that triggers an error, while just hasTooEarlyFormat
should not. Compare hasUnconfirmedVariousArtists vs hasVariousArtists.