MBS-11507: Remove useless resend-verification code
The whole resend-verification code was added in 2011 (MBS-3188)
apparently under the incorrect assumption that we store the email
a user has entered before it gets verified. In fact, we only store it
as part of the verification process - so there's no possible case
where we would have an email stored that still needs verifying.
The entirety of the resend-verification code is as such
completely useless, since there's no reason to re-verify
an already verified email address (even though that's what
the Reverify test was testing).