MBS-9851: Allow finding edits w/ edit notes by not original editor
It can be very useful to find out edits where discussion
of some sort is happening.
This operator will never apply to any other Role::User predicates
than EditNoteAuthor, but I didn't find a way to add it only
for that one - it's still only exposed for that one though.