MBS-11520: Convert Remove Relationship Type edit to React
This adds some more data we were storing, but not displaying
(link phrases, attributes). Some of the phrases are missing
in older edits, as usual, and a few times the attribute itself
might have been removed.
In general, changed this to use display_data and work in a
similar way as AddRelationshipType. Since these always auto-apply,
there's no point linking to the type itself like in AddRelationshipType,
as it will be gone.