MBS-11519: Convert Edit Relationship Type edit to React
This changes the format to be a bit more similar to AddRelationshipType.

Many pre-NGS edits are a huge mess since they store an ID for the
relationship type that now maps to some other different relationship
type post-NGS. This adds a warning if the entity types for the reltype
do not match the ones stored when the edit was added, which hopefully
will add the warning to most of the broken edits.