MBS-11546: Allow Resident Advisor /exchange URLs for releases
RA has over 500 entries on their RA Exchange podcast series,
under /exchange (we have three of them in the DB right now).
These shouldn't be blocked, and should autoselect to show notes.
Since we don't support autoselecting to two different types
for the same entity, this excludes them from the main RA cleanup
and adds a separate cleanup entry for them, to be merged with MBS-9902.