MBS-11548: Show recent additions voted by auto-editors
Require an implicit review by auto-editors through a yes vote or a
direct approval of the added cover art to display recent additions on
the front page.

Approvals are too rare to refresh recent additions often enough. That’s
why it also checks for yes votes (not necessarily the last vote though).