MBS-11558: Show current tracklist data on removal edits
This shows the current tracklist (and a warning) if there are
significant changes between the current tracklist for the medium
being removed and the tracklist that got stored when entering
the removal edit. These are only shown if the edit is open, since
otherwise there's probably no real use to them.