MBS-11457: Drop the series ordering_attribute column
This was added back when we thought we would have different
attributes to store item numbers for part of series relationships.
We don't, so this is always one and the same, 788 (number).

This changes the views to hardcode to 788 and removes the
ordering_attribute column elsewhere. It keeps the
SERIES_ORDERING_ATTRIBUTE constant on JS since that seems to be
used in fields and just points to the 'number' attribute,
but removes it on Perl because it was only used to store
the ID under ordering_attribute and never use it again, AFAICT.