MBS-11564: Don't block Niconi Commons links with Nicovideo cleanup
commons.nicovideo links are separate from the general
nicovideo system and should not be autoselected as streamingfree
(if they should be autoselected at all then probably as downloadfree,
but that is an issue for a future PR). They also shouldn't be blocked,
which is what the cleanup code was currently doing.
For now, this patch just makes the system ignore them so that
they can be entered as needed.