MBS-11569: Limit entity types when creating collection from sidebar
When creating a collection from an entity sidebar, we allowed
selecting any collection type for it, even though an entity
(of a specific entity type that might not match the collection type)
was going to be added to it anyway.
This adds two blocks to that behaviour. First, it loads only
the type values that match the entity type passed to the form.
Additionally, it only adds the entity that was seeded if the entity type
of the final created collection is the same as the one originally passed
to make sure there's no possible case we've missed where an entity
would still get added to a collection of the wrong entity type.