MBS-9631: Add 1200px to available image sizes
Seems the IA has triggered a reindex to get all 1200px
thumbnails generated, so it seems like a good time to add this.
I expect we still want large-thumbnail, not huge-thumbnail,
in root/static/scripts/common/coverart.js for the thumbnail URL,
but added it everywhere else we have thumbnails - including to data
as huge-thumbnail in case userscripts or whatnot want to use it.