MBS-1658 (impl): Add collection comment to collection lists
This allows setting and displaying collection comments from the
main page for a collection. While we will eventually probably want
to also allow setting the comment while adding the entity to the
collection, this seems like a pretty useful start.

Since the comment is just added via ws/js, it should be easy
to also call that endpoint on addition once we come up
with a sensible UI to do so.