MBS-1588: Better errors for invalid CD TOC
When a user tries to attach an invalid CD TOC, just telling them
that it is invalid isn't very useful. They probably didn't generate it
themselves and have no idea what to do with this info. If they even report it
then they're likely to report it to MBS, but it's most likely an issue
with the software they used to generate the CD TOC in the first place.
This makes the error message hopefully more useful.
Also, the /move action also had the same error message,
but no CD TOC is provided to the action to begin with,
just a (Medium) CD TOC id. Changed the error message
to specify that the id is what is invalid.