Fix Catalyst warning: don't try using undef to access hash
Entities that haven't been created yet and are only being
previewed lack both gid and id, but we were trying to use
the result of gid_or_id to access a $loaded entry.
That will never work and results in a warning, so this just
makes sure to only do that if gid_or_id has a result.