Fix catalyst warning: don't try to trim/sanitize undef "strings"
This causes warnings once it gets passed to NFC. If $t is empty,
it seems to be turned into '' during sanitizing, so we might as well
just return '' to begin with and avoid warnings.

This is triggered at least by _clean_link_attribute with no credit
or text value, but looking at logs, I think it's triggered elsewhere too.