MBS-11500: Do not pass GID of removed release to avoid linking (#2035)
It's probably better not to pass the GID here (even though we have it
stored) because the presence of a GID will cause the release to be
displayed with a link that goes nowhere (since the release has since
been removed).

Taking the GID off from the release passed to Artwork would also break
the display of artwork and filename for removed releases. Given
that the artwork for a removed release should have been removed itself
(CAA-126), this is probably not a big issue, but just in case this
passes a separate release entity with a GID to Artwork.

There's also no reason that I can see to form the filename on the
edit itself rather than using artwork.filename, so I changed that
to no longer depend on the release GID on display_data.