MBS-10711: Convert report RecordingRelationshipList to react-table
The previous implementation hid the recording artist
in the RecordingsWithFutureDates report. The recording artist
is an important field for editors, and that report isn't that crowded
that it would need to lose a column. As such, this adds it back
and reduces the amount of conditionals on the component.

This used to be ordered with the recording artist before the recording
title / link. This is something we do not do anywhere else, and we don't
even consistently do it for reports (releases list artist after title),
so I'm changing it to be consistent with other recording displays.