MBS-11612: Restore link to genre/delete to genre sidebar
MBS-10830 stopped showing removal links on the sidebar
unless the entity can be deleted, but genres didn't have
a can_delete method. I could just hardcode genres to always
pass can_delete true or skip the can_delete check for genres,
but in case we decide to add blocks to deletion later,
maybe it's better to just do this like for other entities.