Ensure CAA item ownership during upload process
An exploit exists whereby someone can submit cover art edits that
reference items on archive.org outside of CAA control.

To resolve this, we (1) make sure we're the owner of the bucket in
/ws/js/cover-art-upload before returning a policy that can be used for
uploads, and (2) block all edit submissions that haven't gone through
/ws/js/cover-art-upload first.

(2) is accomplished by returning a nonce that must be provided with the
edit submission.

I made some additions to contrib/ssssss.psgi that helped me test the
changes locally.

Thanks to Ruben Opdebeeck (@ROpdebee) for identifying this issue,
researching and suggesting the solution used here, and reviewing the
code.