MBS-10711: Convert report ReleaseGroupList to react-table
There's only one release group report, as far as I can tell,
that uses keys for special display. Since that's less trivial
to convert to react-table, this keeps the old style list
for that report alone, while converting the general list to react-table.

This used to be ordered with the RG artist before the RG
title / link. This is something we do not do anywhere else, and we don't
even consistently do it for reports (releases list artist after title),
so I'm changing it to be consistent with other RG displays.