MBS-10711: Use ReleaseList for ReleasesSameBarcode report
This used to have three columns, Barcode/ Release / RG.
It seems fine to me to make it have four, Barcode / RG /
Release / (Release) Artist, to follow other release reports. Moving RG
to be before, not after Release because it seems important
to keep the two together for ease of comparison.
Adding the artist seems also useful to see whether the two releases
are likely to be the same (if the artist doesn't match at all,
it's more likely to be an error).