MBS-11619: Ignore periods and +tags in e-mail admin searches
These are often used by spammers to get "new" emails
for more accounts, since many providers (such as Gmail)
allow to add periods and/or +tags to the address to create
aliases for the same account. We don't want to fully block this
(many people use email+musicbrainz@provider to receive
their MB emails, for example) but it should be easy to find people
using it for nefarious purposes.

This is intentionaly untranslatable since it's admin-only.