MBS-11609: Show genre disambiguation on autocomplete dropdown
Rather than just showing the value as the label, this defines
specific labels for the genres - just the name in most cases,
but when there's a disambiguation comment we show it too.
It won't be actually added to the tag list since it's not part
of the value.

jQuery Autocomplete doesn't really support HTML labels by default,
took the code for that from
https://stackoverflow.com/questions/3488016/using-html-in-jquery-ui-autocomplete
which seems to just take it from an extension last updated in 2013.
Given that fact, we might as well just skip the extension.