Add slave-only triggers for *_first_release_date tables
We previously modified our existing triggers to keep this table
up-to-date in production. However, triggers aren't created on slaves, so
these tables would quickly become stale there (affecting features on the
website and webservice) without some sort of solution.

This commit introduces the concept of slave-only triggers that will be
created on slaves. These will never modify replicated tables. (See the
comment on top of CreateSlaveOnlyTriggers.sql for more info.)