MBS-11639: Actually detect attribute type changes
ko.subscribe runs after the value has changed, so this.typeID
is by then always the same as newTypeID, making the previous
implementation of this useless. We need to capture the previous
ID with 'beforeChange' to be able to then compare with it.

When testing this I realized the original intended implementation
was too strict - it would have blanked a pasted free text value
if the user misclicked and selected the wrong free text type,
then corrected their choice. As such, this now only blanks the type
when at least one of the old and new attribute types
does not support free text values.