Add a script to test schema change upgrade scripts
I haven't hooked this into the automated tests yet, but it can be
run manually for now. After configuring the databases as explained in
the comment, you can run it like:

  SUPERUSER=michael REPLICATION_TYPE=2 ./t/script/CheckSchemaMigration.sh

where SUPERUSER is probably 'postgres' if you're on Linux, and
REPLICATION_TYPE matches the numeric constants in
lib/MusicBrainz/Server/Replication.pm.