MBS-1801 (I): SQL additions to support in-site social aspects
This adds five tables to their own messaging namespace.
message is for user-to-user messages - as such, it has a sender
and a receiver, plus a title and content. It also has an optional
parent column which will store the previous message when there
is a message thread.

hidden_message allows for either the sender or receiver (or both)
to "remove" a message from their messaging list. This is less complex
than removing the message outright (on demand from which of the users?)
and makes sure that the content is available to an admin to verify abuse
without the user needing to keep the abusive message in their inbox.

notification is for system notifications. These could be implemented
as messages from ModBot, but having them in a separate table
seems simpler since we are likely to want to display them separately.
It also makes it easier to deal with them separately if we decide to,
for example, make notifications be removed some time after being read.
In preparation for such a possibility, "read" is a timestamp
for notification, but just a boolean for message where we expect
to only remove (hide) messages on request by the sender or receiver.

edit_thanks and edit_note_thanks allow editors to thank others
for their edits and edit notes, respectively. The thanking itself
is stored in these tables, and I expect it should be possible for a user
to check all their sent or received thanks later on.
The user being thanked should also be notified, so on a new thanking
a notification should be inserted in the notification table.
Whether this will happen via a trigger or through code
can be decided later.