MBS-11643: Don't disable links that haven't ended yet (#2093)
We used to disable any link marked as ended, even if the end date
was in the future. This doesn't make much sense, so I made it so
isDisabledLink actually checks for the end date too.
This also uses isDisabledLink, rather than just ended, to decide
whether to display a link on the sidebar. I think it makes sense
that we would also explicitly stop showing any links marked with
isGreyedOut on the sidebar, so it seems like a win-win.