MBS-11631: Unblock label and place links in DNB validation (#2080)
Checking the existing links in the DB, places seem to look the same
as artists. Labels often do too, but there's a different path too,
dnbn, which might be associated with a label code? In any case,
it's valid and we have a fair amount in the DB, so we should support it.