MBS-11630: Do not break on empty edits array (#2085)
There is at least one case where the edits.releaseGroup array
is empty: when the release group has been renamed elsewhere
after the editor opened this release editor session to the same name
the editor is trying to apply here. As such, we try to access edit_type
of an undefined edit. This just uses optional chaining to avoid
that attempted call entirely.