BuildMaterializedTables: use separate transactions (#2111)
Right now, if one of the tables fails to build for some reason, you have
to repeat the whole script again (instead of just specifying the ones
that failed). This can be quite an annoyance, as I recently found out in
the prod migration, because the tables take a long time to build.

This builds each table in a separate transaction, and defaults to the
MAINTENANCE database to guarantee session-related settings
(statement_timeout).