MBS-11679: Wish editors a happy MB anniversary
This uses a similar system as the birthday banner, but it shows
on the editor's MusicBrainz registration anniversary
(also known as "Cake Day" by Reddit people).
Since member_since is not a partial date, this is calculated
in a somewhat different, more complicated way, but it seems
to work just fine. It seems interesting to tell the editor
how many years they have been with MB - I expect some will be surprised
about how long it has been!