Autocomplete2: Generate items list
Instead of managing the list of items to be displayed manually -- which
is tedious and error-prone, especially when we have to splice items out
of the list when clearing the recent entities -- use a `generateItems`
function to calculate the complete list whenever an update is required,
ensuring they're always in the order we want.

We also use a similar technique to generate the status message. This
makes the menu state a bit more declarative and easier to maintain.