Autocomplete2: Tweak menu style
The orange border around the menu doesn't provide enough contrast IMO,
so I'm changing it to a darker gray and increasing the width on the
bottom/right sides to add some depth.

`max-width: 150%` can produce some pretty wide menus if the input itself
is very wide, so I'm capping it at 350px which looks reasonable on the
two systems I tried.