Autocomplete2: Pass "state" object as prop
Previously, the autocomplete "state" (which, given that this is a
stateless component, are technically props -- but I'm using "state" to
refer to anything potentially updated by the reducer) were all lumped
into the props object. This commit moves the "state" into props.state,
so that the props look like this:

  {children, dispatch, state}

This more closely matches how the component will be used in practice.
The dispatch method will come from a higher-level component, not be part
of the state object. So, the state object no longer has to be spread in
JSX (see the updated tests for what this means).

I believe this also makes memoization more efficient on Autocomplete2,
because there are less top-level props to compare, so I've enabled that.