Autocomplete2: Don't show menu if empty
Prior to this commit, if you typed something, we'd show an empty menu
(visible only as a gray line) until the results loaded. This was due to
some broken logic related to `highlightedIndex`. I believe the idea was
that the menu should be opened if we typed something and there was a
visible item to show. However, `getFirstHighlightableIndex` is
calculated based on `items`, which hadn't been updated yet at the point
where this was called.