MBS-11371: Group members in JSON-LD by date, do not squash
The code used to squash all "member of" rels between a group
and a person into one relationship for JSON-LD, whether that made sense
or not (for example, it'd happily squash two relationships into
"start date 2020, end date 2010").
We talked to Google people and they said they thought
having one role entry per date set was better, so this code
does that. We still group several relationships from the same date
(to display "guitar, piano" together if the same member played both
with the same dates).