Disable @babel/plugin-transform-react-constant-elements
Unfortunately, this plugin is still buggy and caches elements that
shouldn't be.  I spent another hour debugging issues due to this.  The
one I'd previously found (https://github.com/mwiencek/babel-issue-11686)
was fixed, but the latest version has other issues I can't be bothered
to report.  It's safer to avoid this plugin at all costs.

Note that if a React element is truly constant, we can just move it to
an outer scope ourselves!  Nothing this plugin does is magic.