Improve formatSetlist performance
This is currently slower than it needs to be because we're parsing the
text twice: once to convert the setlist into an HTML string, and again
to convert the HTML string into React elements (via `expand2react`).

Instead, we can convert the setlist directly into React elements.

Running each version of the function 1000x on the setlist from
https://musicbrainz.org/event/b54fb46e-4e92-4c6c-a171-a19cc5fea8b7 shows
the original to take 0.3108367890119553 seconds and the new version to
take 0.05375143200159073. The output is identical other than a trailing
`<br />` from the new one, though that matches the original Perl version.