MBS-10997: Convert /search/error files to React (#1620)
This is a straight conversion for the most part. I changed the string
on NoResults, because using the result of get_plural_type
in the middle of the sentence just won't work in any languages
where there's also changes by case and not just singular/plural changes.
It would 100% break in Estonian, at least.

Also removing search/form.tt which is now unused.