MBS-11290: Report of works of type X part of another work of type X (#1834)
There's plenty of cases where new users add every movement
of a concerto (or sonata, or whatever) with the same type
as the parent work.  That's pretty much never correct.

This shouldn't have many (if any) false positives from popular music,
since part of work relationships are very uncommon there,
and even then it seems unlikely that "Song is part of Song" is correct.
We could always disable this for "Song" if needed, but that would also
hide correct hits where someone has, say, selected "Song"
for a parent "Song-Cycle" work so I'd rather avoid it
unless users complain.