MBS-11682: Fix deadlocks in apply_artist_*_pending_updates (#2130)
`TRUNCATE` requires an `ACCESS EXCLUSIVE` lock on the target table. This
means it will block on any concurrent operation on the table (which, for
the `artist_*_pending_update` tables, means `INSERT` and other
`TRUNCATE` statements).

A deadlock may occur as follows:

  1. Txn 1 does `INSERT INTO artist_release_group_pending_update` (e.g.
     in the `a_del_release` trigger).
  2. Txn 2 does `INSERT INTO artist_release_group_pending_update`.
  3. Txn 2 does `TRUNCATE artist_release_group_pending_update` in the
     deferred "apply" trigger. This blocks on the `INSERT` from txn 1.
  4. Txn 1 does `TRUNCATE artist_release_group_pending_update` in the
     deferred "apply" trigger. This blocks on txn 2.
  5. Deadlock.

I'm resolving this by using `DELETE` instead, which I believe works
because in this case it only blocks concurrent deletions on the same
rows and not concurrent insertions.