MBS-11709: Beta: Edit notes with { show raw HTML
`expand2react` is intended to expand our i18n syntax, so if it
encounters a `{`, it expects it to be the start of a variable
substitution. Edit notes are not translated strings; the HTML we're
given from the server should be displayed as-is, so we should be using
`dangerouslySetInnerHTML` here.