On release pages, load 100 tracks or 10 mediums
If a release has more than 10 mediums, we currently load it completely
collapsed.  This modifies that behavior to always load up to 100 tracks
or 10 mediums, whichever is hit first.

One side effect of this is that we may now preload some discs even if
the release has more than 10 mediums (as long as the sum of the track
counts on those discs is under 100).

The more important effect, though, is that we don't preload
single-medium releases that have more than 100 tracks. This is half of
MBS-11703; the other half will be to page the tracks.