Return a Map from groupBy/keyBy
I'd like to remove the `__proto__: empty` hacks which are buggy (see
Flow issue 8685).  We're currently returning objects with null
prototypes from groupBy/keyBy, because these objects are meant to be
used as maps.  So, we should really just be using maps.  I made that
change, which actually simplifies some of the usages.