Ensure tooltip content has standard formatting
Tooltips placed inside an h1 inherit the h1 formatting,
which does not seem desirable. This ensures that the content
of a tooltip is formatted in the same way as standard text
on the website by default.

I found about this because the tooltip for ArtistUsageCreditLink
got h1 formatting when used on the artist-credit/ page header.