MBS-6811: Make edit note permalink link to edit page (#2136)
The current implementation just adds an anchor to the same page
the user clicks the permalink from. As such, for any paginated pages,
the permalink is anything but - it will stop working as soon
as the edit slips to the next page.
While this loses the option to link to an edit in a particular context,
this context would always be a moving target in risk of disappearing
except for very carefully curated edit search results that won't change
in time. As such, it seems like a bigger benefit to just have a permalink
that always works.
This adds target="_blank" to ensure that it won't navigate away
from someone's edit list without them being able to vote / submit notes,
but drops it when on the edit page itself since there it makes more sense
to just highlight the note and jump to the anchor directly
like in the current implementation.