MBS-11764: Make user tags sortable by count
I don't see a benefit of reverse alphabetical sorting, so this
only adds sorting on counts (both directions).
The button is kept inline for cases where we only need
one of the two (currently, that's always the votes selector,
but this should also work if we want only the other one).

We currently do the sort in Perl - I considered moving all sorting
to the DB level with ORDER BY but I wasn't sure if that would be
an improvement, so I added these to Perl too.