MBS-11772: Add Greek version to featured artist reports
Not adding this to guessFeat for now since I'm not sure
whether it is gramatically correct without the parens
and the full version seems to take different forms of the Greek "with",
so it doesn't seem trivial to move to the AC. It still makes sense
to provide it on the reports for people who do speak Greek.

For case-insensitive checking to work, this needs to use a different
collation than the default C one, hence 'COLLATE musicbrainz'.