MBS-11776: Readd "loading-message" class to the medium loading message
We used to have this before the React conversion, and it is used
by some userscripts which don't really have a great substitute.
This used to be in a div inside the td, but that div didn't seem
to have any use other than it being easy to append with JQuery.
Hopefully just having it on the td itself works as well.