MBS-11796: Add Internet Archive logo for sidebar
For some reason, we were still using "no-favicon" for the IA.
It seems sensible to use this for archive.org links. It might sometimes
also show for Wayback Machine links in addition to direct links
(given that AFAICT this just picks the last match it finds in the URL,
I guess it will do that for any Wayback Machine links that do not already
match another favicon). I don't see that as a negative at all though.