MBS-11804: Allow filtering edit search by time of edit note
We now have two different predicates that require natural date
processing, so this moves the relevant sections to a role.

I also moved the cardinality map to the role - Date was reusing
the ID predicate operators, but that is confusing (it's not that
the two are connected in any way, other than just happening to use
the same operators). As such, it seems sensible to just
stop making Date depend on ID. Other than moving the operators
to the role, that makes combine_with_query larger but, IMO,
a lot clearer.