MBS-11806: Don't group relationships for different track sets
As far as I can tell, this code was written for relationships
that apply to one entity only (one recording, one release...).
As such, there was never a need to check for more than
"are the rels related, and are the attributes the same".
But this now also used for grouped display of medium relationships
under a release, and that involves relationships from different tracks.
Nothing in the code checked whether the tracks actually
matched before they got combined - one of the sets of tracks
would just get dropped when combining the two relationship
groups. This now makes sure that the tracks also match, if they exist,
before allowing a merged display of the relationship.