MBS-11811: Do not batch-change track data for collapsed media
Even when a user has closed (collapsed) a certain medium in the
release editor, running a batch-change action such as global guess case
or changing all matching artist credits still changes the track data
for that medium. This can be very confusing, since the editor
does not see the changes at all unless they either reopen the medium
or pay close attention to the edit previews on the Edit Note tab.
As such, this changes the way those calls work so that
they will ignore any medium that is collapsed at the time
the change is made.
This was already the case for media which were never expanded
in the first place, although that's only because there were no loaded
tracks to modify. This only added to the confusion, since there is
no good way to see at a glance whether a medium has been loaded
and then closed or has never been loaded.