MBS-11833: Drop "f." from the featured artists reports
Almost all cases I can see are false positives: stuff like "(F. Chopin),
or just "F. Whatever remix". It seems to be pretty uncommon for f. X
to actually be used as feat, so while it could make sense
to support it on the guess feat button (where it would be used
mostly if the user already knows there's featured artists
on a tracklist or title), the false positives make it unwise to have it
on the reports.
For what it's worth, we don't seem to even use this in featRegex
for the button either.