MBS-11854: Recognize unicode hyphen in guess case
Words with a unicode hyphen were not being recognized
as split words: the hyphen was just being considered one more character.
This changes that, and then ensures that guess case recognizes
the hyphen as such and does the same as with a hyphen-minus.

Sadly, part of my fix for MBS-10156 (for re‐mode with unicode hyphen)
will no longer work now since it is no longer considered
one single word, but I think this makes a lot more sense for now.
The current code already didn't do anything with re-mode with a
hyphen-minus, so this actually adds consistency.