Test loading DBDefs module before starting server
Occasionally a server instance would continuously error, we assume it
likely did not start with a valid `lib/DBDefs.pm`. In production,
this file is generated from `docker/templatesDBDefs.pm.ctmpl` and a JSON
config file using git2consul/consul-agent/consul-template.

It is not clear which of these components is the cause of this problem.
The content of the `lib/DBDefs.pm` is not known either, but it isn’t
missing or empty. (Different error messsages would be thrown then.)

This patch just tries loading DBDefs module before running the server.
If the module cannot be loaded, consul-template will try starting a new
instance automatically, hopefully using a correct configuration provided
by consul-agent in the meantime. It assumes to fix a potential bug with
consul-template not restarting the server when erroring continuously.

Disclaimer: This does NOT test anything beyond loading DBDefs module such as:
validity of configuration values, connection to the database or any other
service.

Impl. note: Server won’t reload `lib/DBDefs.pm` if it changed, so there is no
need to check DBDefs all the time while running. Consul is in charge of
starting/stopping/restarting the server.