MBS-11846: Display release artist on release group view
Right now, there's no way to see if a release inside a release group
has a different artist credit than the release group itself.
This adds a release artist column to that view.
It also adds an "artist-credit-variation" class for cases
where the release and release group artist credits differ,
to help userscript users highlight that (matching what we do
for track artists on recording pages).